intelliherd/requirements.txt
aniso8601==8.0.0
Babel==2.7.0
blinker==1.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Ext==0.1
flask-jsontools==0.1.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-paypal==0.1.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
paypalrestsdk==1.11.7
psycopg2==2.8.4
pycparser==2.19
pyOpenSSL==19.1.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
SQLAlchemy-serializer==1.3.4.2
stripe==2.40.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
